Rename chat_serverice to supervisor
