+++ src/backend/functions/testing_chat.py
@@ -2,7 +2,7 @@ from services.outlook_agent import outlook_agent
 from services.todo_agent import todo_agent
 import azure.functions as func
 from helper.azure_config import AzureConfig
-from services.chat_service import Supervisor
+from services.superivsor import Supervisor
 from typing import Dict
 from langchain.agents import AgentExecutor
 from langchain_core.load import dumps
