+++ src/backend/services/superivsor.py
@@ -1,42 +1,48 @@
-from typing import Dict
+from datetime import datetime
+from typing import Dict, List
 import functools
 import operator
 from helper.azure_config import AzureConfig
 from langchain.agents import AgentExecutor
 from langchain.output_parsers.openai_functions import JsonOutputFunctionsParser
-from langchain_core.messages import BaseMessage, HumanMessage
+from langchain_core.messages import AIMessage, BaseMessage, HumanMessage
 from langchain_core.prompts import ChatPromptTemplate, MessagesPlaceholder
 from langchain_core.prompts.chat import ChatPromptTemplate, MessagesPlaceholder
 from langchain_openai import AzureChatOpenAI
 from langgraph.graph import StateGraph, END
 from typing import Annotated, Sequence, TypedDict
+import uuid
+
+# from langchain_community.adapters import convert
+# from openai import ChatCompletion
 
 # https://langchain-ai.github.io/langgraph/tutorials/multi_agent/agent_supervisor/#construct-graph
 # https://langchain-ai.github.io/langgraph/how-tos/pass-run-time-values-to-tools
 
-class Supervisor:
-    def __init__(self, config: AzureConfig, system_prompt: str, agents: Dict[str, AgentExecutor]):
 
-        #credential = DefaultAzureCredential()
+class Supervisor:
+    def __init__(
+        self, config: AzureConfig, system_prompt: str, agents: Dict[str, AgentExecutor]
+    ):
+
+        # credential = DefaultAzureCredential()
 
         llm = AzureChatOpenAI(
             deployment_name=config.azure_deployment,
             openai_api_version=config.azure_openai_api_version,
             # azure_endpoint=config.azure_endpoint,
-            
             # todo: mi / env var
             # azure_ad_token_provider=credential.get_token
-
             # error:
             #   'DefaultAzureCredential failed to retrieve a token from the included
-            #    credentials.\nAttempted credentials:\n\tEnvironmentCredential: 
+            #    credentials.\nAttempted credentials:\n\tEnvironmentCredential:
             #    EnvironmentCredential authentication unavailable. Environment variables
-            #    are not fully configured.\nVisit 
-            #    https://aka.ms/azsdk/python/identity/environmentcredential/troubleshoot 
+            #    are not fully configured.\nVisit
+            #    https://aka.ms/azsdk/python/identity/environmentcredential/troubleshoot
             #    to troubleshoot this issue.\n\tManagedIdentityCredential: "get_token" requires
             #    at least one scope\nTo mitigate this issue, please refer to the troubleshooting
-            #    guidelines here at 
-            #    https://aka.ms/azsdk/python/identity/defaultazurecredential/troubleshoot., 
+            #    guidelines here at
+            #    https://aka.ms/azsdk/python/identity/defaultazurecredential/troubleshoot.,
             #    NoneType: None'
         )
 
@@ -93,23 +99,19 @@ class Supervisor:
         workflow = StateGraph(AgentState)
 
         for agent_name, agent in agents.items():
-            node = functools.partial(
-                self.__agent_node__,
-                agent=agent,
-                name=agent_name
-            )
+            node = functools.partial(self.__agent_node__, agent=agent, name=agent_name)
             workflow.add_node(agent_name, node)
 
         for member in members:
             # We want our workers to ALWAYS "report back" to the supervisor when done
             workflow.add_edge(member, "supervisor")
-        
+
         # The supervisor populates the "next" field in the graph state
         # which routes to a node or finishes
         # our conditional map is the agents and the finish state
         conditional_map = {k: k for k in members}
         conditional_map["FINISH"] = END
-        # is there the possibility and value in having additiona conditional edges? what would be their src 
+        # is there the possibility and value in having additiona conditional edges? what would be their src
         workflow.add_conditional_edges(
             "supervisor", self.__should_continue__, conditional_map
         )
@@ -120,21 +122,23 @@ class Supervisor:
 
     def __agent_node__(self, state, agent, name):
         result = agent.invoke(state)
-        return {"messages": [HumanMessage(content=result["output"], name=name)]}
+        return {"messages": [AIMessage(content=result["output"], name=name, id=str(uuid.uuid4()), date=str(datetime.now().isoformat()))]}
 
     def __should_continue__(self, state, config):
         print(state)
         print(config)
-        
+
         # the value of the next field is the key in the conditional map
         # if our conditional map is just a dictionary of string, string
         # why can't this function just return the value of the dictionary element?
         # is there some value to that layer of abstraction?
-        return state["next"] 
+        return state["next"]
 
+    async def arun(self, in_messages: List[BaseMessage]) -> List[BaseMessage]:  # is that typing right?
+        state = self.graph.invoke(in_messages)
+        out_messages: List[BaseMessage] = state["messages"]
+        return out_messages
 
-    def run(self, messages):  # is that typing right?
-        return self.graph.invoke(messages)
 
 # The agent state is the input to each node in the graph
 class AgentState(TypedDict):
